Fix import-side type issues exposed by postgres backend
Four bugs surfaced when running `--import` against the new postgres
backend. Sqlite was permissive enough to mask each one; postgres is
not. Fixed in place — sqlite and mysql users see no behaviour change.

- Device.serial_number is declared `Integer` but Garmin's Descent Mk3i
  family reports serials > 2**31 (e.g. 3476167836). The hard-coded
  fallback `unknown_device_serial_number = 9999999999` would also
  overflow on a strict 32-bit backend. Switch to BigInteger on the
  primary key and on the matching FK columns in DeviceInfo / File and
  on activities.device_serial_number / activity_laps.device_serial_number.

- monitoring_info.file_id was `Integer`, but the value passed in is the
  string id from `File.id` (which itself is `String`). Sqlite quietly
  coerced "269351985837" to a numeric value; postgres rejects the
  string-to-int4 cast outright. Switch monitoring_info.file_id to
  `String` so the FK shape matches `File.id`.

- monitoring_fit_file_processor.__unpack_tuple only unpacked when
  type(value) is tuple, but fitfile yields per-activity-type arrays as
  plain `list`. The list bled through to a `Float` column and the
  failure poisoned the SQLAlchemy session for every subsequent insert
  in the same file. Accept both list and tuple via isinstance().

+++ garmindb/garmindb/activities_db.py
@@ -6,7 +6,7 @@ __license__ = "GPL"
 
 import logging
 import datetime
-from sqlalchemy import Column, String, Float, Integer, Boolean, DateTime, Time, Enum, ForeignKey, PrimaryKeyConstraint, desc, literal_column
+from sqlalchemy import Column, String, Float, Integer, BigInteger, Boolean, DateTime, Time, Enum, ForeignKey, PrimaryKeyConstraint, desc, literal_column
 from sqlalchemy.ext.declarative import declared_attr
 from sqlalchemy.orm import relationship
 from sqlalchemy.ext.hybrid import hybrid_property
@@ -109,7 +109,7 @@ class Activities(ActivitiesDb.Base, ActivitiesCommon):
     sport = Column(String)
     sub_sport = Column(String)
 
-    device_serial_number = Column(Integer)
+    device_serial_number = Column(BigInteger)
 
     self_eval_feel = Column(String)
     self_eval_effort = Column(String)
@@ -326,7 +326,7 @@ class ActivitiesDevices(ActivitiesDb.Base, idbutils.DbObject):
     table_version = 1
 
     activity_id = Column(String)
-    device_serial_number = Column(Integer)
+    device_serial_number = Column(BigInteger)
     __table_args__ = (PrimaryKeyConstraint("activity_id", "device_serial_number"),)
 
     @classmethod
